Tweaking preQ "education level" section formatting
I think it's quite a bit easier to read now.  Radio buttons are aligned
with the left edge of the page better thanks to the use of padding-top
and padding-bottom.

Also, now the text fields asking for focus/major are properly sitting to
the right of their respective radio buttons, rather than just wrapping
under them.  This increases readability substantially, I feel.