Transition from training, to game, to program end
Game starts in training mode (i.e. it reads from the training config
file and does not log data) and displays appropriate descriptive text in
the MainFrame.  Following training mode completion, the game
automatically transitions to the full 8-minute game (the MainFrame
descriptive text is changed appropriately, the full game config file is
read in, and data is logged).  Following completion of the full game,
the program prints a confirmation message to the console and then closes
itself.