Working on changing patient terms
Changing "Patient A" to "Standard Patient" and "Patient B" to "Surgery
Patient".

Changed the text that appears on the "Assign Patient" Buttons, and also
the labels in the player's waiting room panel.

Also, changed the title of the waiting room from "Waiting Room" to
"Patient Waiting Room", which I think increases clarity.