Properly-named CSVs now save to ResearchData\UID
- CSV files now save to ResearchData\UID and are named "UID_P.csv" and
"UID_A.csv" for the player and agent files repectively ("UID" is a
stand-in for the actual UID)
- Remove trialNum and userId fields from class MainFrame; they are no
longer needed
- Left in "condition" var in class MainFrame, not sure if needed; it is
not currently being used for anything.  Should check w/ sponsor.

Note: Currently the program will fail to write to the log file if the
directory ResearchData\UID does not exist.
Any existing files with the same names will be overwritten.